+++ man/roxygen/meta.R
@@ -3,6 +3,7 @@ list(
     admin = "Other administrative units datasets: ",
     stats = "Other statistical units datasets: ",
     transport = "Other transport networks datasets: ",
-    services = "Other basic services datasets: "
+    services = "Other basic services datasets: ",
+    extra = "Additional utils for downloading datasets: "
   )
 )